requirements.txt
streamlit
pandas
matplotlib
mysql-connector-python
openai
python-dotenv
openpyxl
tabulate
sqlalchemy 
psycopg2 
plotly 
seaborn

